app/requirements.txt
pika
yagmail
python-dotenv